requirements.txt
openai>=1.0.0
fastmcp>=2.0.0
httpx>=0.24.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
gradio>=4.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
